refactor: move src/components tests into test/src/components
Move all 29 component test files from src/components/*.spec.ts to
test/src/components/ and update import paths accordingly.
