+++ test/src/components/CommandCell.spec.ts
