+++ test/src/components/FilesOverlay.spec.ts
