+++ test/src/components/LauncherCell.spec.ts
