+++ test/src/components/TerminalCell.spec.ts
