+++ test/src/components/TerminalGrid.spec.ts
