+++ test/src/components/TimelineOverlay.spec.ts
