+++ test/src/components/ToolsPane.spec.ts
