+++ test/src/components/cellFlip.spec.ts
