+++ test/src/components/cmEditor.spec.ts
