+++ test/src/components/dropPaths.spec.ts
