+++ test/src/components/skillSeed.spec.ts
