+++ test/src/components/terminalViewActive.spec.ts
